pyproject.toml
[build-system]
requires      = ["hatchling"]
build-backend = "hatchling.build"

[project]
name            = "houseflip"
version         = "0.1.0"
description     = "Real estate listing scraper and price opportunity analyzer"
requires-python = ">=3.13"
dependencies = [
    "httpx>=0.27",
    "parsel>=1.9",
    "pydantic>=2.7",
    "duckdb>=1.0",
    "polars>=1.0",
    "streamlit>=1.36",
    "tenacity>=8.3",
    "rich>=13.0",
    "altair>=5.3",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-httpx>=0.30",
    "ruff>=0.4",
    "mypy>=1.10",
]

[tool.hatch.build.targets.wheel]
packages = ["src/houseflip"]

[tool.ruff]
target-version = "py313"
line-length    = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]

[tool.mypy]
python_version = "3.13"
strict         = true
plugins        = ["pydantic.mypy"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths    = ["tests"]
